% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{aptitude_nonpara}
\alias{aptitude_nonpara}
\title{aptitude_nonpara Function}
\usage{
aptitude_nonpara(p, npop)
}
\arguments{
\item{p}{A numeric vector representing percentiles of talent values.}

\item{npop}{A numeric value representing the population size.}
}
\value{
A numeric vector of talent values extracted from percentiles.
}
\description{
This function extracts talent values from percentiles.
}
\details{
The aptitude_nonpara function takes a vector of percentiles \code{p}, along with optional parameter \code{npop},
and extracts talent values based on a non-parametric method.

The function first converts order statistics to their percentiles using the \code{order_pbino} function.
Then, it transforms these percentiles to Gaussian values corresponding to a general population of a greater than or equal size.

The parameter \code{npop} represents the population size, which is used in the transformation process.
}
\keyword{percentiles}
\keyword{talent,}
