% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligiblepop.R
\docType{data}
\name{eligiblepop}
\alias{eligiblepop}
\title{Historical Eligible Dataset}
\format{
A tibble with 5481 observations of 2 variables:
\itemize{
\item year: a numeric vector representing the year of a historical event.
Negative values may indicate years BCE (Before Common Era)/ Before Christ
\item cumulativepop: a numeric vector representing the cumulative population count or
total number of individuals considered in the specific year
}
}
\usage{
data(eligiblepop)
}
\description{
This dataset contains information on historical events, including the year of occurrence
and the population count for each specific year.
}
\keyword{datasets}
