% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{era_adjust_nonpara}
\alias{era_adjust_nonpara}
\title{era_adjust_nonpara}
\usage{
era_adjust_nonpara(
  anchor_talent,
  talents_environ,
  ytilde_environ,
  npop_environ
)
}
\arguments{
\item{anchor_talent}{Numeric. The latent talent value that we want to map onto the common mapping environment.}

\item{talents_environ}{Numeric vector. Latent talent values in the common mapping environment.}

\item{ytilde_environ}{Numeric vector. The surrogate points Ytilde defining the interpolated CDF of the common mapping environment.}

\item{npop_environ}{Numeric. The talent pool size in the common mapping environment.}
}
\value{
Numeric value of the era-adjusted statistic
}
\description{
Maps a given latent talent to a common mapping environment
and gets the era-adjusted statistic using non-parametric approach
}
\details{
Computes era-adjusted statistic using equation (2.7) from Yan et al. (2025, AOAS)
}
