% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{k_finder}
\alias{k_finder}
\title{k_finder Function}
\usage{
k_finder(
  x,
  method = "robust",
  outlier_handling = "FALSE",
  stab = 0.01,
  cutoff = 0.014
)
}
\arguments{
\item{x}{A numeric vector representing the input data}

\item{method}{A string indicating use robust version (default) vs. original/legacy version}

\item{outlier_handling}{A logical value; if TRUE, increases k to amplify the influence of extreme values so that potential outliers are more pronounced.}

\item{stab}{A numeric value representing the stability parameter}

\item{cutoff}{A numeric value representing the cutoff when deciding the tail.}
}
\value{
A numeric vector containing the optimal value of k, along with values K1 and K2
}
\description{
This function finds the optimal value of k based on input data
}
\details{
The k_finder function computes the optimal value of k based on the input vector \code{x} and a stability parameter \code{stab}.
It performs linear and quadratic modeling to evaluate the optimal value of k that maximizes the explained variance in the data.

The function initializes quantities for linear approximation and then evaluates various values of k within specific ranges.
It uses a combination of linear and quadratic models to estimate the optimal value of k
}
\keyword{approximation,}
\keyword{linear}
\keyword{modeling,}
\keyword{optimization,}
\keyword{order}
\keyword{statistics}
