% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{talent_computing_nonpara}
\alias{talent_computing_nonpara}
\title{talent_computing_nonpara}
\usage{
talent_computing_nonpara(ystarstar, y, npop)
}
\arguments{
\item{ystarstar}{Numeric. Additional parameter for talent estimation.}

\item{y}{Numeric vector. Observed statistic, arranged from highest to lowest.}

\item{npop}{Numeric. Population size.}
}
\value{
Numeric vector of estimated talent values.
}
\description{
Estimate underlying talent values using a non-parametric method.
}
\details{
Computes talent values for \code{y} using \code{Ftilde()} and \code{aptitude_nonpara()}.
}
