% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talentpool.R
\docType{data}
\name{talentpool}
\alias{talentpool}
\title{MLB Talent Pool}
\format{
A tibble with 155 observations of 3 variables:
\itemize{
\item year: a numeric vector for year.
\item NLpop: a numeric vector for the talent pool size for the NL.
\item ALpop: a numeric vector for the talent pool size for the AL.
}
}
\source{
See \href{https://eckeraadjustment.web.illinois.edu/MLB-talent-pool-v3.0.html}{this technical report}
for details on how these estimates of the talent pool were constructed from
1871 to 2025.
}
\usage{
data(talentpool)
}
\description{
This dataset contains the talent pool for the American League (AL) and
National League (NL) from 1871 to 2025.
}
\keyword{datasets}
