% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{Aptitude_nonpara}
\alias{Aptitude_nonpara}
\title{Aptitude_nonpara Function}
\usage{
Aptitude_nonpara(p, alpha = 1.16, npop)
}
\arguments{
\item{p}{A numeric vector representing percentiles of talent values.}

\item{alpha}{A numeric value representing the shape parameter of the Pareto distribution. Default is 1.16.}

\item{npop}{A numeric value representing the population size.}
}
\value{
A numeric vector of talent values extracted from percentiles.
}
\description{
This function extracts talent values from percentiles.
}
\details{
The Aptitude_nonpara function takes a vector of percentiles \code{p}, along with optional parameters \code{alpha} and \code{npop},
and extracts talent values based on a non-parametric method.

The function first converfs order statistics to their percentiles using the \code{order_pbino} function.
Then, it transforms these percentiles to Pareto values corresponding to a general population of a greater than or equal size.

The parameter \code{alpha} represents the shape parameter of the Pareto distribution, with a default value of 1.16.
The parameter \code{npop} represents the population size, which is used in the transformation process.
}
\keyword{percentiles}
\keyword{talent,}
