% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batters_adjusted.R
\docType{data}
\name{batters_adjusted}
\alias{batters_adjusted}
\title{Batters Adjusted Dataset}
\format{
A data frame with \link{number_of_rows} rows and \link{number_of_columns} variables:
\itemize{
\item\link{name}:  a character vector containing the name of the player.
\item\link{playerID}: The unique ID of the player.
\item\link{age}: The age of the player during the season (e.g., integer).
\item\link{year}: The season year (e.g., integer).
\item\link{PA}: Plate appearances, the total number of times the player has appeared at the plate (e.g., integer).
\item\link{AB}: At-bats, the number of official at-bats the player has taken (e.g., integer).
\item\link{H}: Hits, the total number of successful hits by the player (e.g., integer).
\item\link{HR}: Home runs, the number of home runs hit by the player (e.g., integer).
\item\link{BB}: Bases on Balls (Walks), the number of times the player was walked (e.g., integer).
\item\link{BA}: Batting Average, calculated as H/AB (e.g., numeric).
\item\link{OBP}: On-Base Percentage, calculated as (H + BB + HBP)/(AB + BB + HBP + SF) (e.g., numeric).
\item\link{HBP}: Hit by Pitch, the number of times the player was hit by a pitched ball (e.g., integer).
\item\link{SF}: Sacrifice Fly, a fly ball that allows a runner to score after the catch, but does not count as an at-bat (e.g., integer).
\item\link{ebWAR}: Era-adjusted Wins Above Replacement as computed by Baseball Reference (e.g., numeric).
\item\link{efWAR}: Era-adjusted Wins Above Replacement as computed by FanGraphs (e.g., numeric).
}
}
\usage{
data(batters_adjusted)
}
\description{
A dataset containing adjusted statistics for baseball batters.
}
\keyword{datasets}
