% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{compute_ystarstar}
\alias{compute_ystarstar}
\title{compute_ystarstar Function}
\usage{
compute_ystarstar(x, k, stab = 1e-04)
}
\arguments{
\item{x}{A numeric value representing the input data.}

\item{k}{A numeric value representing the parameter k}

\item{stab}{A numeric value representing the stability parameter. Default is 0.0001.}
}
\value{
A list containing the computed optimal value of ystarstar, along with auxiliary information.
}
\description{
This function computes the optimal value of ystarstar based on input data and a tail
probability modeling algorithm described in the Details section.
}
\details{
With a model \eqn{h} selected (more on this below), we find \eqn{Y_i^{**}} as the solution
of the following optimization problem
\deqn{
Y_i^{**} = \text{argmin}_y|h^{-1}(Y_{i,(n_i)}) - \tilde{F}_{Y_i}(Y_{i,(n_i)};y)|,
}
where \eqn{\tilde{F}_{Y_i}(\cdot;y)} is \eqn{\tilde{F}_{Y_i}} with \eqn{y}
replacing \eqn{Y_i^{**}} in its construction.

We select a tail probability model via the following steps:
\enumerate{
\item Fit the logistic and logistic quadratic models. Use BIC to choose among these models.
\item Stop if the above optimization problem can be solved.
\item If the above optimization problem cannot be solved, then fit the log-logistic and log-logistic quadratic models. Use BIC to choose among these models.
\item Stop if the above optimization problem  can be solved.
\item If the above optimization problem cannot be solved, then fit the logistic cubic model.
}
}
\keyword{approximation,}
\keyword{linear}
\keyword{modeling,}
\keyword{optimization,}
\keyword{order}
\keyword{statistics}
