% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{compute_ystarstar}
\alias{compute_ystarstar}
\title{compute_ystarstar Function}
\usage{
compute_ystarstar(x, k, stab = 1e-04)
}
\arguments{
\item{x}{A numeric value representing the input data.}

\item{k}{A numeric value representing the parameter k}

\item{stab}{A numeric value representing the stability parameter. Default is 0.0001.}
}
\value{
A list containing the computed optimal value of ystarstar, along with auxiliary information.
}
\description{
This function computes the optimal value of ystarstar based on input data
}
\details{
The compute_ystarstar function computes the optimal value of ystarstar based on the input vector \code{x},
parameter \code{k}, and stability parameter \code{stab}

The function performs various linear and quadratic modeling techniques to estimate the optimal value of ystarstar
that maximizes the explained variance in the data
}
