% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{Ftilde}
\alias{Ftilde}
\title{Ftilde Function}
\usage{
Ftilde(y, t, ystar)
}
\arguments{
\item{y}{A numeric vector representing the input data.}

\item{t}{A numeric value indicating the target value.}

\item{ystar}{A numeric value representing an additional parameter.}
}
\value{
A transformed value computed based on the input parameters.
}
\description{
This function computes a transformed value based on the input parameters.
}
\details{
The Ftilde function computes a transformed value based on the input data vector \code{y}, a target value \code{t}, and an additional parameter \code{ystar}.

The algorithm sorts the input vector \code{y} and computes a new vector \code{ytilde} using a specific formula.
Then, it computes the transformed value based on the relationship between \code{t} and the elements of \code{ytilde}.

If \code{t} is greater than or equal to the maximum value in \code{ytilde}, the function returns a constant value.
If \code{t} is less than or equal to the minimum value in \code{ytilde}, the function returns 0.
Otherwise, it computes the transformed value based on the position of \code{t} relative to the elements of \code{ytilde}.
}
\examples{
# Example usage of the Ftilde function
y <- c(1, 2, 3, 4, 5)
t <- 3.5
ystar <- 1
Ftilde(y, t, ystar)

}
\keyword{algorithm,}
\keyword{transformation,}
\keyword{vector}
