% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{k_finder}
\alias{k_finder}
\title{k_finder Function}
\usage{
k_finder(x, stab = 1e-04)
}
\arguments{
\item{x}{A numeric vector representing the input data}

\item{stab}{A numeric value representing the stability parameter}
}
\value{
A numeric vector containing the optimal value of k, along with values K1 and K2
}
\description{
This function finds the optimal value of k based on input data
}
\details{
The k_finder function computes the optimal value of k based on the input vector \code{x} and a stability parameter \code{stab}.
It performs linear and quadratic modeling to evaluate the optimal value of k that maximizes the explained variance in the data.

The function initializes quantities for linear approximation and then evaluates various values of k within specific ranges.
It uses a combination of linear and quadratic models to estimate the optimal value of k
}
\keyword{approximation}
\keyword{linear}
\keyword{modeling,}
\keyword{optimal}
\keyword{value,}
