% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchers_adjusted.R
\docType{data}
\name{pitchers_adjusted}
\alias{pitchers_adjusted}
\title{Era-adjusted Pitchers Dataset}
\format{
A data frame with 30932 rows and 10 variables:
\itemize{
\item name:  a character vector containing the name of the player.
\item playerID: The unique ID of the player.
\item age: The age of the player during the season (e.g., integer).
\item year: The season year (e.g., integer).
\item IP: Innings pitched, the total number of innings the pitcher has thrown (e.g., numeric).
\item ERA: Earned Run Average, calculated as (9 * ER)/IP (e.g., numeric).
\item SO: Strikeouts, the number of batters struck out by the pitcher (e.g., integer).
\item ebWAR: Era-adjusted Wins Above Replacement as computed by Baseball Reference (e.g., numeric).
\item efWAR: Era-adjusted Wins Above Replacement as computed by FanGraphs (e.g., numeric).
\item ER: Earned Runs, the number of runs scored off the pitcher that are not the result of errors or passed balls (e.g., integer).
}
}
\usage{
data(pitchers_adjusted)
}
\description{
A dataset containing era-adjusted statistics for pitchers.
}
\details{
These era-adjusted statistics are obtained from Full House Modeling.
This model computes era-adjusted statistics through a principled
balancing of how well a player performed "vs their peers" and the
size of the MLB talent pool. Under this model, great all-time
statistics requires that an MLB player is both better than their
peers and played during a time in which the talent pool is large.
In this way, the model constructs an even playing field that
extends across eras.
}
\references{
Shen Yan, Adrian Burgos Jr., Christopher Kinson, and Daniel J. Eck (2024). "Comparing baseball players across eras via novel Full House Modeling." Available at: \url{https://arxiv.org/abs/2207.11332}

Website: \url{https://eckeraadjustment.web.illinois.edu/}
}
\keyword{datasets}
