% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchers_adjusted.R
\docType{data}
\name{pitchers_adjusted}
\alias{pitchers_adjusted}
\title{Pitchers Adjusted Dataset}
\format{
A data frame with \link{number_of_rows} rows and \link{number_of_columns} variables:
\itemize{
\item\link{name}:  a character vector containing the name of the player.
\item\link{playerID}: The unique ID of the player.
\item\link{age}: The age of the player during the season (e.g., integer).
\item\link{year}: The season year (e.g., integer).
\item\link{IP}: Innings pitched, the total number of innings the pitcher has thrown (e.g., numeric).
\item\link{ERA}: Earned Run Average, calculated as (9 * ER)/IP (e.g., numeric).
\item\link{SO}: Strikeouts, the number of batters struck out by the pitcher (e.g., integer).
\item\link{ebWAR}: Era-adjusted Wins Above Replacement as computed by Baseball Reference (e.g., numeric).
\item\link{efWAR}: Era-adjusted Wins Above Replacement as computed by FanGraphs (e.g., numeric).
\item\link{ER}: Earned Runs, the number of runs scored off the pitcher that are not the result of errors or passed balls (e.g., integer).
}
}
\usage{
data(pitchers_adjusted)
}
\description{
A dataset containing adjusted statistics for baseball pitchers.
}
\keyword{datasets}
