% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{talent_computing_nonpara}
\alias{talent_computing_nonpara}
\title{talent_computing_nonpara Function}
\usage{
talent_computing_nonpara(ystar, data, npop, alpha = 1.16)
}
\arguments{
\item{ystar}{A numeric value representing an additional parameter for talent estimation.}

\item{data}{A data frame containing the input data.}

\item{npop}{A numeric value representing the population size.}

\item{alpha}{A numeric value representing the shape parameter of the Pareto distribution. Default is 1.16.}
}
\value{
A data frame containing estimated talent values based on the input data.
}
\description{
This function estimates underlying talent values using a non-parametric method.
}
\details{
The talent_computing_nonpara function estimates underlying talent values using a non-parametric approach based on the input data.

It computes talent values for the input data \code{data$hpi} using the Ftilde and Aptitude_nonpara functions.
The parameter \code{ystar} is an additional parameter used in talent estimation.
}
\keyword{Pareto}
\keyword{distribution}
\keyword{estimation,}
\keyword{non-parametric,}
\keyword{talent}
